console/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.27.0
email-validator>=2.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth>=2.25.0
